requirements.txt
streamlit>=1.30.0
soundfile>=0.12.1
librosa>=0.10.1
numpy>=1.24.0
ffmpeg-python>=0.2.0
torch>=2.2.0
faster-whisper>=1.0.3
demucs>=4.0.0
nvidia-cublas-cu12>=12.6.0; platform_system == "Linux"
nvidia-cudnn-cu12>=9.5.0; platform_system == "Linux"
